backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.0.7
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
jwcrypto==1.5.6
numpy==2.0.1
oauthlib==3.2.2
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
